requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
cssselect==1.2.0
discord==2.3.2
discord.py==2.3.2
fake-useragent==1.5.1
frozenlist==1.4.1
idna==3.7
importlib_metadata==7.1.0
lxml==5.2.2
lxml_html_clean==0.1.1
multidict==6.0.5
parse==1.20.2
pyee==11.1.0
pyppeteer==2.0.0
pyquery==2.0.0
python-dotenv==1.0.1
requests==2.32.3
requests-html==0.10.0
soupsieve==2.5
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==1.26.19
w3lib==2.2.1
websockets==10.4
yarl==1.9.4
zipp==3.19.2
